Converter/two_level_morphanalyzer/requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
backports.zoneinfo==0.2.1
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1693583441880/work
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
click==8.1.7
cmake==3.27.4.1
contourpy==1.1.0
cycler==0.11.0
dataclasses==0.6
datasets==2.6.1
decorator==5.1.1
dill==0.3.5.1
Django==4.1.7
django-ckeditor==6.5.1
django-debug-toolbar==4.2.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-sslserver==0.22
djangorestframework==3.14.0
evaluate==0.3.0
exceptiongroup==1.1.3
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gast==0.3.3
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.43.2
gradio_client==0.5.0
grpcio==1.58.0
h11==0.14.0
h5py==2.10.0
hparams==0.3.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
lit==16.0.6
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.13
networkx==3.1
numba==0.57.1
numpy==1.20.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pooch==1.7.0
protobuf==3.20.0
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pyTelegramBotAPI==4.13.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.2.0
referencing==0.30.2
regex==2023.8.8
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
responses==0.18.0
rpds-py==0.10.2
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.5.0
semantic-version==2.10.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.6
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
telebot==0.0.5
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-io-gcs-filesystem==0.34.0
threadpoolctl==3.2.0
tokenizers==0.13.3
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision @ file:///croot/torchvision_1689078161692/work
tqdm==4.66.1
transformers==4.33.1
triton==2.0.0
typeguard==4.1.5
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
